+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/TypeSpec.kt
@@ -171,7 +171,7 @@ public class TypeSpec private constructor(
         codeWriter.emitAnnotations(annotations, false)
         codeWriter.emitModifiers(
           modifiers,
-          if (isNestedExternal) setOf(PUBLIC, EXTERNAL) else setOf(PUBLIC),
+          implicitModifiers + if (isNestedExternal) setOf(PUBLIC, EXTERNAL) else setOf(PUBLIC),
         )
         codeWriter.emit(kind.declarationKeyword)
         if (name != null) {
@@ -446,6 +446,7 @@ public class TypeSpec private constructor(
         ANNOTATION in modifiers -> emptySet()
         EXPECT in modifiers -> setOf(EXPECT)
         EXTERNAL in modifiers -> setOf(EXTERNAL)
+        INTERNAL in modifiers -> setOf(INTERNAL)
         else -> emptySet()
       }
     }
@@ -454,6 +455,7 @@ public class TypeSpec private constructor(
       return defaultImplicitFunctionModifiers + when {
         EXPECT in modifiers -> setOf(EXPECT)
         EXTERNAL in modifiers -> setOf(EXTERNAL)
+        INTERNAL in modifiers -> setOf(INTERNAL)
         else -> emptySet()
       }
     }
@@ -462,6 +464,7 @@ public class TypeSpec private constructor(
       return defaultImplicitTypeModifiers + when {
         EXPECT in modifiers -> setOf(EXPECT)
         EXTERNAL in modifiers -> setOf(EXTERNAL)
+        INTERNAL in modifiers -> setOf(INTERNAL)
         else -> emptySet()
       }
     }